+++ src/mplchart/primitives/__init__.py
@@ -1,16 +1,17 @@
 """mplchart drawing primitives"""
 
+from .candlesticks import Candlesticks
 from .axes import NewAxes, SameAxes
 from .price import Price
 from .volume import Volume
 from .ohlc import OHLC
-from .candlesticks import Candlesticks
 from .peaks import Peaks
 from .stripes import Stripes
 from .markers import Markers
 from .lineplot import LinePlot
 from .areaplot import AreaPlot
 from .barplot import BarPlot
+from .zigzag import ZigZag