+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
 ## 0.0.18
+- Added `ZigZag` Primitive (Experimental)
 - `Price` primitive now accepts calculation items like 'hlc', 'hlcc', ...
 - `Price` can now be used and composed as an indicator like in ```EMA(20) @ Price('hlcc')```